Split things up so we've got better separation
