+++ server/src/main/java/com/octopus/teamcity/opentelemetry/server/helpers/HelperPerBuildOTELHelperFactory.java
@@ -1,20 +1,12 @@
 package com.octopus.teamcity.opentelemetry.server.helpers;
 
-import com.octopus.teamcity.opentelemetry.server.LogMasker;
-import com.octopus.teamcity.opentelemetry.server.OTELService;
-import io.opentelemetry.exporter.otlp.trace.OtlpGrpcSpanExporter;
-import io.opentelemetry.exporter.otlp.trace.OtlpGrpcSpanExporterBuilder;
-import io.opentelemetry.exporter.zipkin.ZipkinSpanExporter;
+import com.octopus.teamcity.opentelemetry.server.endpoints.OTELEndpointFactory;
 import io.opentelemetry.sdk.trace.SpanProcessor;
-import io.opentelemetry.sdk.trace.export.BatchSpanProcessor;
-import io.opentelemetry.sdk.trace.export.SpanExporter;
 import jetbrains.buildServer.serverSide.BuildPromotion;
 import jetbrains.buildServer.serverSide.ProjectManager;
-import jetbrains.buildServer.serverSide.crypt.EncryptUtil;
 import org.apache.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 import static com.octopus.teamcity.opentelemetry.common.PluginConstants.*;
@@ -23,11 +15,15 @@ public class HelperPerBuildOTELHelperFactory implements OTELHelperFactory {
     static Logger LOG = Logger.getLogger(HelperPerBuildOTELHelperFactory.class.getName());
     private final ConcurrentHashMap<Long, OTELHelper> otelHelpers;
     private final ProjectManager projectManager;
+    @NotNull
+    private final OTELEndpointFactory otelEndpointFactory;
 
     public HelperPerBuildOTELHelperFactory(
-        ProjectManager projectManager
+        ProjectManager projectManager,
+        @NotNull OTELEndpointFactory otelEndpointFactory
     ) {
         this.projectManager = projectManager;
+        this.otelEndpointFactory = otelEndpointFactory;
         LOG.debug("Creating HelperPerBuildOTELHelperFactory.");
 
         this.otelHelpers = new ConcurrentHashMap<>();
@@ -49,25 +45,10 @@ public class HelperPerBuildOTELHelperFactory implements OTELHelperFactory {
                 if (params.get(PROPERTY_KEY_ENABLED).equals("true")) {
                     var endpoint = params.get(PROPERTY_KEY_ENDPOINT);
                     SpanProcessor spanProcessor;
-                    if (params.get(PROPERTY_KEY_SERVICE).equals(OTELService.HONEYCOMB.getValue())) {
-                        Map<String, String> headers = new HashMap<>();
-                        headers.put("x-honeycomb-dataset", params.get(PROPERTY_KEY_HONEYCOMB_DATASET));
-                        headers.put("x-honeycomb-team", EncryptUtil.unscramble(params.get(PROPERTY_KEY_HONEYCOMB_APIKEY)));
-                        spanProcessor = buildGrpcSpanProcessor(headers, endpoint);
-                    } else if (params.get(PROPERTY_KEY_SERVICE).equals(OTELService.ZIPKIN.getValue())) {
-                        spanProcessor = buildZipkinSpanProcessor(endpoint);
-                    } else {
-                        Map<String, String> headers = new HashMap<>();
-                        params.forEach((k, v) -> {
-                            if (k.startsWith(PROPERTY_KEY_HEADERS)) {
-                                var name = k.substring(PROPERTY_KEY_HEADERS.length());
-                                name = name.substring(1, name.length() - 1);
-                                var value = EncryptUtil.isScrambled(v) ? EncryptUtil.unscramble(v) : v;
-                                headers.put(name, value);
-                            }
-                        });
-                        spanProcessor = buildGrpcSpanProcessor(headers, endpoint);
-                    }
+
+                    var otelHandler = otelEndpointFactory.getOTELEndpointHandler(params.get(PROPERTY_KEY_SERVICE));
+                    spanProcessor = otelHandler.BuildSpanProcessor(endpoint, params);
+
                     long startTime = System.nanoTime();
                     var otelHelper = new OTELHelperImpl(spanProcessor);
                     long endTime = System.nanoTime();
@@ -81,28 +62,6 @@ public class HelperPerBuildOTELHelperFactory implements OTELHelperFactory {
             LOG.debug(String.format("Using NullOTELHelper for build %d.", buildId));
             return new NullOTELHelperImpl();
         });
-    }
-
-    private SpanProcessor buildZipkinSpanProcessor(String exporterEndpoint) {
-        String endpoint = String.format("%s/api/v2/spans", exporterEndpoint);
-        ZipkinSpanExporter zipkinExporter = ZipkinSpanExporter.builder()
-                .setEndpoint(endpoint)
-                .build();
-
-        return BatchSpanProcessor.builder(zipkinExporter).build();
-    }
-
-    private SpanProcessor buildGrpcSpanProcessor(Map<String, String> headers, String exporterEndpoint) {
-
-        OtlpGrpcSpanExporterBuilder spanExporterBuilder = OtlpGrpcSpanExporter.builder();
-        headers.forEach(spanExporterBuilder::addHeader);
-        spanExporterBuilder.setEndpoint(exporterEndpoint);
-        SpanExporter spanExporter = spanExporterBuilder.build();
-
-        LOG.debug("OTEL_PLUGIN: Opentelemetry export headers: " + LogMasker.mask(headers.toString()));
-        LOG.debug("OTEL_PLUGIN: Opentelemetry export endpoint: " + exporterEndpoint);
-
-        return BatchSpanProcessor.builder(spanExporter).build();
     }
 
     @Override
