+++ server/src/main/resources/META-INF/build-server-plugin-Octopus.TeamCity.Opentelemetry.xml
@@ -10,5 +10,6 @@
     <bean class="com.octopus.teamcity.opentelemetry.server.ProjectConfigurationSettingsController"/>
     <bean class="com.octopus.teamcity.opentelemetry.server.BuildOverviewExtensionController"/>
     <bean class="com.octopus.teamcity.opentelemetry.server.helpers.HelperPerBuildOTELHelperFactory"/>
+    <bean class="com.octopus.teamcity.opentelemetry.server.endpoints.OTELEndpointFactory"/>
     <bean class="com.octopus.teamcity.opentelemetry.server.BuildStorageManagerImpl"/>
 </beans>
